Skip monmap update if monmap does not exist

+++ src/mon/start_mon.sh
@@ -16,13 +16,16 @@ function start_mon {
     exit 1
   fi
 
-  ceph-monstore-tool /mon/data get monmap -- --out /tmp/monmap
-  monmaptool --clobber --rm $MON_ID "/tmp/monmap"
-  monmaptool --clobber --add $MON_ID $IP:6789 "/tmp/monmap"
+  monmap_exists=$(ceph-monstore-tool /mon/data get monmap -- --out /tmp/monmap > /dev/null)$? || true
 
-  /usr/bin/ceph-mon "${DAEMON_OPTS[@]}" -i "${MON_ID}" --inject-monmap /tmp/monmap --mon-data "$MON_DATA_DIR" --public-addr $IP
+  # If monmap exists, update our ip
+  if [[ $monmap_exists -eq 0 ]] ; then
+    monmaptool --clobber --rm $MON_ID "/tmp/monmap"
+    monmaptool --clobber --add $MON_ID $IP:6789 "/tmp/monmap"
+    /usr/bin/ceph-mon "${DAEMON_OPTS[@]}" -i "${MON_ID}" --inject-monmap /tmp/monmap --mon-data "$MON_DATA_DIR" --public-addr $IP
+  fi
 
-    # start MON
+  # start MON
   log "Starting Ceph-Mon"
   #while true; do sleep 2; done
   exec /usr/bin/ceph-mon "${DAEMON_OPTS[@]}" -i "${MON_ID}" --mon-data "$MON_DATA_DIR" --public-addr $IP
